agent/requirements.txt
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.10.0
